Finishes org level logic
